+++ ClosestStringProblem/instance.hpp
@@ -16,24 +16,24 @@
 
 class Instance {
     FILE *instance_file; // File with the instance specification
-    int k; // alphabet size
-    int n; // number of strings
-    int l; // string length
+    long int k; // alphabet size
+    long int n; // number of strings
+    long int l; // string length
     char * alphabet; // k characters
     char ** strings; // n strings
-    int *** char_to_string;
-    int ** strings_per_char_count;
+    long int *** char_to_string;
+    long int ** strings_per_char_count;
     
 public:
     Instance(const char * csp_file_name);
     ~Instance();
-    int getAlphabetSize() {return k;};
-    int getNumberOfStrings() {return n;};
-    int getStringLength() {return l;};
+    long int getAlphabetSize() {return k;};
+    long int getNumberOfStrings() {return n;};
+    long int getStringLength() {return l;};
     char * getAlphabet() {return alphabet;};
     char ** getStrings() {return strings;};
-    int ** getStringsPerCharCount() {return strings_per_char_count;};
-    int *** getCharToString() {return char_to_string;};
+    long int ** getStringsPerCharCount() {return strings_per_char_count;};
+    long int *** getCharToString() {return char_to_string;};
     void print();
 };
